+++ dashboard-web/src/App.tsx
@@ -4,6 +4,7 @@ import { StatsOverview } from './dashboard/StatsOverview'
 import { SelectorList } from './dashboard/SelectorList'
 import { SelectorDetail } from './dashboard/SelectorDetail'
 import { ChronicSelectors } from './dashboard/ChronicSelectors'
+import { EfficacyDashboard } from './dashboard/EfficacyDashboard'
 
 export default function App() {
   return (
@@ -14,6 +15,7 @@ export default function App() {
           <Route path="/selectors" element={<SelectorList />} />
           <Route path="/selectors/:id" element={<SelectorDetail />} />
           <Route path="/chronic" element={<ChronicSelectors />} />
+          <Route path="/efficacy" element={<EfficacyDashboard />} />
         </Route>
       </Routes>
     </BrowserRouter>